Fetch and shove DB_POOL into DATABASE_URL string
See this issue:

  https://github.com/janko-m/sinatra-activerecord/issues/56

For now it looks like it's either the database.yml file, OR the
DATABASE_URL env variable -- not the combination of the two, which is
disappointing.

Until something can be written that merges the rest of the config file
into the connection string this hacky hack will do. I can at least set
the DB POOL in here now.